+++ packages/confirmation-modal/src/ConfirmationModal.story.tsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import { ComponentStory } from '@storybook/react';
 import Button from '@leafygreen-ui/button';
 import ConfirmationModal, { Variant } from '.';
+import defaultArgTypes from '../../../stories/defaultArgTypes';
 
 export default {
   title: 'Packages/Modals/ConfirmationModal',
@@ -9,9 +10,15 @@ export default {
   args: {
     title: 'Confirm Title Here',
     buttonText: 'Confirm',
-    requiredInputText: false,
     children:
       'This is some description text, and it is extra long so it fills up this modal. Another thing about the modals here. This is some description text, and it is extra long so it fills up this modal. Another thing about the modals here.',
+  },
+  argTypes: {
+    requiredInputText: { control: 'text' },
+    open: { control: 'boolean' },
+    title: { control: 'text' },
+    buttonText: { control: 'text' },
+    children: defaultArgTypes.children,
   },
 };